wrapper: work around an ld bug
ld.bfd used to have a bug which stemmed from using a bfd representing a
real file to hold phony sections. It manifested when ld was used with
lto-plugin and was fixed in 6cd255ca1f03550291bd05ac4548e383bca88c5f
(sourceware.org/ml/binutils/2016-04/msg00353.html). However this fix is
not enough for our case since (unlike lto-plugin) our plug-priv can
claim crt{1,i,begin}.o files. In this way the phony sections get created
in a real dso's bfd and confuse the linker when it later tries to handle
them.

We work around this by always passing a dummy object file, which doesn't
get claimed by the plugin, to the linker. This file precedes all other
input files on the command line so it becomes the victim of bfd linker's
black magic (it does not suffer though because it is a dummy).

The dummy file should have appropriate elfclass (and have some proper
sections but no symbols) for the system on which the software is used.
Currently we achieve this by extracting it from libg.a (from glibc-devel
package) already built for the platform.