util/gcc-wrapper.c: decline flto flag on the second run.
Object files do not goes through all plugins, so, if lto
plugin reads lto object file and produce a regular object file,
we do not see the produced object file. If our plugin is first
then we obtain lto object file and do not know what to do with it.