force/deps-general: add some_variable
This should help gcc's configure be identical on 1 and 2 runs.

The problem without this fix:
Some other package has conftest.c with the same hash (which is easy:
"int some_variable = 0;") which it compiles *and links*. And the
corresponding conftest in gcc suffers even though it is not linked
because the compiler plugin does its thing.